2019.06.29
1. 폴더 구조를 src 폴더 안으로 다 넣음

2. /java_test/src/calculatordemo_1/CalculatorDemo_1.java
/java_test/src/calculatordemo_1/CalculatorDemo_2.java
/java_test/src/calculatordemo_1/CalculatorDemo_3.java
/java_test/src/calculatordemo_1/CalculatorDemo_4.java
파일에 @작성자 : 정현석라고 써놓은 부분은 정현석이 추가적인 코멘트 작성한 부분이니 확인 바람.
