--- input_output/IO_3.java
@@ -1,14 +1,0 @@
-package input_output;
-
-import java.util.Scanner;
-
-public class IO_3 {
-
-	public static void main(String[] args) {
-		 Scanner sc = new Scanner(System.in);
-	        int i = sc.nextInt();
-	        System.out.println(i*1000);
-	        sc.close();
-	}
-
-}
