+++ src/compare/compare.java
