+++ src/input_output/IO_1.java
@@ -1,0 +1,8 @@
+package input_output;
+
+public class IO_1 {
+
+	public static void main(String[] args){
+		System.out.println(args.length);
+		    }
+}
