+++ src/input_output/IO_2.java
@@ -1,0 +1,10 @@
+package input_output;
+
+public class IO_2 {
+
+	 public static void main(String[] args){
+	        for(String e : args){
+	            System.out.println(e);
+	        }
+	    }
+}
